Set up resource parameters assertions generation
