+++ pkg/acceptance/bettertestspoc/assert/resourceparametersassert/gen/templates/definition.tmpl
@@ -1,0 +1,28 @@
+{{- /*gotype: github.com/Snowflake-Labs/terraform-provider-snowflake/pkg/acceptance/bettertestspoc/assert/resourceparametersassert/gen.ResourceParametersAssertionsModel*/ -}}
+
+{{- $assertName := .Name | printf "%sResourceParametersAssert" -}}
+{{- $nameLowerCase := FirstLetterLowercase .Name -}}
+{{- $assertVar := FirstLetter $nameLowerCase }}
+type {{ $assertName }} struct {
+*assert.ResourceAssert
+}
+
+func {{ .Name }}ResourceParameters(t *testing.T, name string) *{{ $assertName }} {
+t.Helper()
+
+{{ $assertVar }} := {{ $assertName }}{
+ResourceAssert: assert.NewResourceAssert(name, "parameters"),
+}
+{{ $assertVar }}.AddAssertion(assert.ValueSet("parameters.#", "1"))
+return &{{ $assertVar }}
+}
+
+func Imported{{ .Name }}ResourceParameters(t *testing.T, id string) *{{ $assertName }} {
+t.Helper()
+
+{{ $assertVar }} := {{ $assertName }}{
+ResourceAssert: assert.NewImportedResourceAssert(id, "imported parameters"),
+}
+{{ $assertVar }}.AddAssertion(assert.ValueSet("parameters.#", "1"))
+return &{{ $assertVar }}
+}
