+++ pkg/acceptance/bettertestspoc/assert/resourceparametersassert/gen/templates/preamble.tmpl
@@ -1,0 +1,16 @@
+{{- /*gotype: github.com/Snowflake-Labs/terraform-provider-snowflake/pkg/acceptance/bettertestspoc/assert/resourceparametersassert/gen.PreambleModel*/ -}}
+
+// Code generated by sdk-to-schema generator; DO NOT EDIT.
+
+package {{ .PackageName }}
+
+import (
+"strconv"
+"testing"
+{{- range .AdditionalStandardImports }}
+"{{- . }}"
+{{- end }}
+
+"github.com/Snowflake-Labs/terraform-provider-snowflake/pkg/acceptance/bettertestspoc/assert"
+"github.com/Snowflake-Labs/terraform-provider-snowflake/pkg/sdk"
+)
