+++ pkg/acceptance/bettertestspoc/assert/resourceparametersassert/generate.go
@@ -1,0 +1,3 @@
+package resourceparametersassert
+
+//go:generate go run ./gen/main/main.go $SF_TF_GENERATOR_ARGS
